Crate hw

use hw;

// Sets display brightness to 60%
let _ = hw::display_brightness(0.6);

// Mutes the systems speaker or connected headphone.
let _ = hw::speaker_unmute();

// Shutdown via a shutdown signal to the host operating system
//let _ = hw::power_shutdown();
 

Crate level functions (display_brightness, speaker_unmute, and power_shutdown) are simple wrappers around their Hw equivalents.

ENVIRONMENT VARIABLES

HWLIB_ERROR_LOG: If set, prints verbose logging of internal events to stderr.

Additional environment variables may be significant on some platforms. refer to the privilege and unprivileged modules for further details.

Structs

Error
Error description type for Hw

Hw
Hardware Handle

PlatformHw
(Linux) Platform-specific Hardware Handle

Enums

ErrorKind
Error Kind

Functions

display_brightness
See Hw

display_off
See Hw

display_on
See Hw

power_off
See Hw

power_reboot
See Hw

power_restart
See Hw

power_shutdown
See Hw

speaker_mute
See Hw

speaker_unmute
See Hw

speaker_volume
See Hw

Type Definitions

Result