Struct hw::Hw

pub struct Hw(_);

Hardware Handle

Methods

impl Hw

fn new() -> Hw

Return a new hardware handle.

fn power_off(&mut self) -> Result<()>

Immediate system power off.

Warning: May cause data loss and other undesirable side-effects associated with an "Unclean" shutdown. Consider System::shutdown if this is undesirable.

fn power_reboot(&mut self) -> Result<()>

Immediate system reboot.

Warning: May cause data loss and other undesirable side-effects associated with an "Unclean" shutdown. Consider System::restart if this is undesirable.

fn power_shutdown(&mut self) -> Result<()>

Sends the shutdown signal for an operating system-assisted power off.

Exact behavior is system-dependent, and may entail additional user action (confirmation). Successful result may not imply successful shutdown initiation.

fn power_restart(&mut self) -> Result<()>

Sends the restart signal for an operating system-assisted reboot.

Exact behavior is system-dependent, and may entail additional user action (confirmation). Successful result may not imply successful shutdown initiation.

fn speaker_volume(&mut self, level: f64) -> Result<()>

Sets system audio output volume to a specified level. level range must be a value either 0.0 (min), 1.0 (max), or any value in between.

fn speaker_mute(&mut self) -> Result<()>

Mutes system audio output

fn speaker_unmute(&mut self) -> Result<()>

Unmutes system audio output

fn display_brightness(&mut self, intensity: f64) -> Result<()>

Sets all system video output brightness to a specified intensity. intensity range must be a value either 0.0 (min), 1.0 (max), or any value in between.

fn display_off(&mut self) -> Result<()>

Turns off all video output for the system.

fn display_on(&mut self) -> Result<()>

Turns on all video output for the system.